# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ZeroAgent Event Venue Marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /venues:
    get:
      operationId: listVenues
      tags: [venues]
      summary: List all approved venues
      parameters:
        - name: city
          in: query
          schema:
            type: string
        - name: minCapacity
          in: query
          schema:
            type: integer
        - name: maxPrice
          in: query
          schema:
            type: number
        - name: eventType
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 12
      responses:
        "200":
          description: List of venues
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VenueListResponse"

    post:
      operationId: createVenue
      tags: [venues]
      summary: Create a new venue listing (owner only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateVenueBody"
      responses:
        "201":
          description: Venue created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Venue"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /venues/{id}:
    get:
      operationId: getVenue
      tags: [venues]
      summary: Get a venue by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Venue details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VenueDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    put:
      operationId: updateVenue
      tags: [venues]
      summary: Update venue (owner only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateVenueBody"
      responses:
        "200":
          description: Updated venue
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Venue"

    delete:
      operationId: deleteVenue
      tags: [venues]
      summary: Delete venue (owner or admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /chat/{bookingId}:
    get:
      operationId: getChatMessages
      tags: [chat]
      summary: Get chat messages for approved booking
      parameters:
        - name: bookingId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatMessage"

    post:
      operationId: sendChatMessage
      tags: [chat]
      summary: Send a message in an approved booking chat
      parameters:
        - name: bookingId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageBody"
      responses:
        "201":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatMessage"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
...
    Venue:
      type: object
      properties:
        id:
          type: integer
        ownerId:
          type: integer
        ownerName:
          type: string
        name:
          type: string
        description:
          type: string
        city:
          type: string
        address:
          type: string
        capacity:
          type: integer
        pricePerDay:
          type: number
        eventTypes:
          type: array
          items:
            type: string
        images:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [pending, approved, rejected]
        createdAt:
          type: string
          format: date-time
      required: [id, ownerId, name, description, city, address, capacity, pricePerDay, eventTypes, images, status, createdAt]
    VenueDetail:
      allOf:
        - $ref: "#/components/schemas/Venue"
        - type: object
          properties:
            bookedDates:
              type: array
              items:
                type: string
                format: date
    VenueListResponse:
      type: object
      properties:
        venues:
          type: array
          items:
            $ref: "#/components/schemas/Venue"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required: [venues, total, page, totalPages]
    CreateVenueBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        city:
          type: string
        address:
          type: string
        capacity:
          type: integer
        pricePerDay:
          type: number
        eventTypes:
          type: array
          items:
            type: string
        images:
          type: array
          items:
            type: string
      required: [name, description, city, address, capacity, pricePerDay, eventTypes]
    UpdateVenueBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        city:
          type: string
        address:
          type: string
        capacity:
          type: integer
        pricePerDay:
          type: number
        eventTypes:
          type: array
          items:
            type: string
        images:
          type: array
          items:
            type: string
...
    ChatMessage:
      type: object
      properties:
        id:
          type: integer
        bookingId:
          type: integer
        senderId:
          type: integer
        senderName:
          type: string
        senderRole:
          type: string
        message:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, bookingId, senderId, senderName, senderRole, message, createdAt]
    SendMessageBody:
      type: object
      properties:
        message:
          type: string
      required: [message]
...