# lib/api-zod/src/generated/types/chatMessage.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ZeroAgent Event Venue Marketplace API
 * OpenAPI spec version: 0.1.0
 */

export interface ChatMessage {
  id: number;
  bookingId: number;
  senderId: number;
  senderName: string;
  senderRole: string;
  message: string;
  createdAt: Date;
}